Chapter 5: Chess Dictionary Validator finished
There are two files of this practice project because I got confused first time through and made a program wherein you input your piece onto a chessboard manually.
The second one is the proper  practice project. Which is supposed to get a dictionary and check it for any discrepancies (wrong positioning, exceeded number of chess pieces and invalid chesspieces)