Chapter 8 finished
This was a pretty straightforward chapter about input validation. It introduced a new module named pyinputplus which makes input validation far less tedious for an average joe programmer such as myself in my current stages. This was further emphasized in the final practice project which forced me to make the chapter project without the use of this new module. Showing me the true convenience in the functions pyinputplus brings.