Chapter 12 finished
Link Verification is done in a miniscule form(for forum posts on a slovenian forum site.). And with that I end the web scraping chapter in Automate the boring stuff.

Offcourse this does not close the chapter on web scraping in my future endeavours. For you see, this chapter, beside my curiosity for the language, was why I bought the book in the first place. I shall continue to brush my interest in web sraping after I have finished this book.

Now onto Working with Excel Spreadsheets!