Chapter 13 finished
The folder is a bit of a mess because I didn't mess with changing the current working directory when making programs.

Other than that this chapter went on for a bit of a long time through the theoretical part for me, this is not blaming the writing ofcourse, just had a few rough days when going into it.

  Going into the practice projects I, for some reason, couldn't  imagine how I could solve them until I started trying to work them through. Here I suprised myself how fast I breezed throgh them. Did them all in two hours.