Chapter 14 complete //Undocumented//
I did the chapter, but omitted the files because they hold the credentials and files which are sensitive.

Working with Google sheets is simple enough. The only thing I was missing throughout the chapter was a way to format cell styles and fonts. Going to have to look into that since it would be very convenient to be able to generate a spreadsheet with templates using python.

Although I realise I might also want to plan a standard template and just make it in sheets. We'll see, still want to do it in Python though.

Onto Chapter 15 now. "Working with PDF and Word documents"!